backend/requirements.txt
amqp==5.2.0
arrow==1.3.0
asgiref==3.7.2
Babel==2.14.0
billiard==4.2.0
binaryornot==0.4.4
celery==5.4.0
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cookiecutter==2.5.0
cron-descriptor==1.4.3
DAWG-Python==0.7.2
Django==5.0
django-autocomplete-light==3.9.7
django-browser-reload==1.12.1
django-celery-beat==2.6.0
django-debug-toolbar==4.2.0
django-localflavor==4.0
django-phonenumber-field==7.3.0
django-tailwind==3.6.0
django-timezone-field==6.1.0
docopt==0.6.2
docxcompose==1.4.0
docxtpl==0.16.8
idna==3.6
Jinja2==3.1.2
kombu==5.3.7
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
num2words==0.5.13
Petrovich==2.0.1
phonenumberslite==8.13.29
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycountry==23.12.11
Pygments==2.17.2
pymorphy3==2.0.1
pymorphy3-dicts-ru==2.4.417150.4580142
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.1.0
python-slugify==8.0.1
python-stdnum==1.19
pytils==0.4.1
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
rich==13.7.0
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
types-python-dateutil==2.8.19.14
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
gunicorn